fix: playtest polish — combate, derrota, UI, Fase-2 e analytics
Gameplay & combate:
- Escala unificada de sprites/personagens
- Dash VFX da Nixie visível no solo
- Investida da Nixie aplicando dano (solo)
- Indicadores de dano menores na HUD

Derrota & câmera:
- Pipeline de morte com foco, vignette e despawn gradual dos inimigos
- Zoom da câmera ajustado ao range do jogador

Conteúdo & fases:
- Arte e colisões do mapa 2 (Fase-2)

UI & fluxo:
- Magículas, barra de vida e preparação (contrato não vem mais selecionado)
- Menu: botão/link de feedback de playtest; bootstrap sem texto na tela
- Pause: reiniciar (solo), quit → menu principal

Áudio & telemetria:
- Crossfade musical alinhado às transições de cena
- GameAnalytics (eventos de gameplay e navegação de telas)

+++ Assets/Scripts/UI/World/DamageIndicatorPresenter.cs
@@ -2,16 +2,13 @@ using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
 
-/// <summary>
 /// Escuta GameEvents.OnDamageShown e instancia números flutuantes de dano.
-/// Coloque na cena (ex.: MultiplayerManagers) ou Bootstrap.
-/// </summary>
 public class DamageIndicatorPresenter : MonoBehaviour
 {
     [SerializeField] private float floatDistance = 0.8f;
     [SerializeField] private float lifetime = 0.75f;
     [SerializeField] private Color damageColor = new Color(1f, 0.35f, 0.25f, 1f);
-    [SerializeField] private int fontSize = 4;
+    [SerializeField] private float fontSize = 4;
 
     private void OnEnable()
     {
@@ -60,9 +57,7 @@ public class DamageIndicatorPresenter : MonoBehaviour
     }
 }
 
-/// <summary>
 /// Anima o número para cima e destrói o objeto.
-/// </summary>
 public class DamageIndicatorFloater : MonoBehaviour
 {
     private float _floatDistance;
